backend/requirements-enterprise.txt
# Enterprise Backend Requirements
# Python 3.11+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
redis[hiredis]==5.0.1
aioredis==2.0.1

# Celery & Task Queue
celery[redis]==5.3.4
flower==2.0.1
kombu==5.3.4

# Authentication & Security
pyjwt==2.8.0
cryptography==41.0.7
python-jose==3.3.0
bcrypt==4.1.1
argon2-cffi==23.1.0

# API Documentation
pydantic==2.5.2
pydantic-settings==2.1.0

# CORS & Security
python-cors==1.0.0
bleach==6.1.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Monitoring & Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==0.42b0
prometheus-client==0.19.0

# AWS Integration
boto3==1.34.11
botocore==1.34.11

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0
factory-boy==3.3.0

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
pre-commit==3.6.0
ipython==8.18.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
aiofiles==23.2.1
python-json-logger==2.0.7
structlog==23.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
orjson==3.9.10

# Email
python-emails==0.6.0
aiosmtplib==3.0.1

# Rate Limiting
slowapi==0.1.9

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Environment Management
python-decouple==3.8

# Compression
python-snappy==0.6.1

# UUID
shortuuid==1.0.11

# Scheduling
apscheduler==3.10.4

# GraphQL (optional)
strawberry-graphql[fastapi]==0.216.1

# Search
elasticsearch==8.11.0

# Machine Learning (optional for AI features)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers==4.36.2
langchain==0.0.350
openai==1.6.1
anthropic==0.8.1
